rename PreAnalysis to AEWTO; split state-access bodies into AbstractStateManager.cpp
Changes:

* PreAnalysis class -> AEWTO; corresponding files renamed.
* Move analysis prep (svfir / icfg / preAnalysis / WTO build) from
  AbstractInterpretation::runOnModule into the AI ctor, so subclass
  ctors can rely on those being ready and do their own per-flavour
  init in the body (cycle ValVar precompute, SVFG build).
* Drop SVFG* / class SVFG forward decl from AbstractInterpretation —
  field moves to FullSparseAbstractInterpretation.
* Constants and the IRGraph::NullPtr sentinel now resolve via the
  trace[Global] fallback in SemiSparseAbstractInterpretation, so the
  type-directed const cases are gone.
* Factor the state-access method bodies (get/has/updateAbsValue,
  state ops, GEP/load/store helpers, def/use queries) for dense /
  semi-sparse / full-sparse into a new AbstractStateManager.cpp.
  Headers unchanged; AI.cpp / SparseAI.cpp keep only driver logic
  (ctor/dtor, runOnModule, transfer functions, cycle helpers,
  buildSVFG).
* Drop the redundant <iomanip> include in AbstractState.cpp; the
  Util/Options.h include re-pulls it transitively.

Build: green under LLVM 18 + LLVM 21.
