+++ svf/lib/AE/Svfexe/AbstractStateManager.cpp
@@ -1,0 +1,545 @@
+//===- AbstractStateManager.cpp -- AE state-access implementations ---//
+//
+//                     SVF: Static Value-Flow Analysis
+//
+// Copyright (C) <2013->  <Yulei Sui>
+//
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
+//===----------------------------------------------------------------------===//
+//
+// State-access bodies factored out of AbstractInterpretation.cpp /
+// SparseAbstractInterpretation.cpp.  Class declarations stay in their
+// respective headers; this file just hosts the implementations of the
+// methods that used to live on the (now-folded) AbstractStateManager —
+// trace lookup, value get/has/update, GEP / load / store helpers, and
+// def/use queries — for dense, semi-sparse, and full-sparse alike.
+//
+
+#include "AE/Svfexe/AbstractInterpretation.h"
+#include "AE/Svfexe/SparseAbstractInterpretation.h"
+#include "SVFIR/SVFIR.h"
+#include "Graphs/SVFG.h"
+#include "Util/Options.h"
+#include "Util/SVFUtil.h"
+
+using namespace SVF;
+
+// =====================================================================
+//  Dense (AbstractInterpretation) — direct trace lookup; sparse
+//  subclasses override the virtuals below.
+// =====================================================================
+
+AbstractState& AbstractInterpretation::getAbsState(const ICFGNode* node)
+{
+    if (abstractTrace.count(node) == 0)
+    {
+        assert(false && "No preAbsTrace for this node");
+        abort();
+    }
+    return abstractTrace[node];
+}
+
+void AbstractInterpretation::updateAbsState(const ICFGNode* node, const AbstractState& state)
+{
+    abstractTrace[node] = state;
+}
+
+void AbstractInterpretation::joinStates(AbstractState& dst, const AbstractState& src)
+{
+    dst.joinWith(src);
+}
+
+bool AbstractInterpretation::hasAbsState(const ICFGNode* node)
+{
+    return abstractTrace.count(node) != 0;
+}
+
+/// Dense base: direct trace lookup, with a top sentinel for genuinely
+/// missing entries (e.g. function parameters like argc, never written
+/// before first read).  Sparse subclasses override with a def-site
+/// resolution chain.
+///
+/// The "in map" check is a raw map.count — NOT inVarToValTable /
+/// inVarToAddrsTable, which gate on isInterval / isAddr.  SVF
+/// canonically represents uninit and null-pointer shapes as
+/// (interval=bottom ∧ addrs=∅); those predicates would falsely report
+/// such an entry as "not present", and the top fallback below would
+/// then clobber the very signal NullptrDerefDetector::isUninit keys off.
+const AbstractValue& AbstractInterpretation::getAbsValue(const ValVar* var, const ICFGNode* node)
+{
+    u32_t id = var->getId();
+    AbstractState& as = abstractTrace[node];
+    if (as.getVarToVal().count(id))
+        return as[id];
+    as[id] = IntervalValue::top();
+    return as[id];
+}
+
+const AbstractValue& AbstractInterpretation::getAbsValue(const ObjVar* var, const ICFGNode* node)
+{
+    AbstractState& as = getAbsState(node);
+    u32_t addr = AbstractState::getVirtualMemAddress(var->getId());
+    return as.load(addr);
+}
+
+const AbstractValue& AbstractInterpretation::getAbsValue(const SVFVar* var, const ICFGNode* node)
+{
+    if (const ObjVar* objVar = SVFUtil::dyn_cast<ObjVar>(var))
+        return getAbsValue(objVar, node);
+    if (const ValVar* valVar = SVFUtil::dyn_cast<ValVar>(var))
+        return getAbsValue(valVar, node);
+    assert(false && "Unknown SVFVar kind");
+    abort();
+}
+
+/// Dense base: direct existence check at `node`.  Mirrors the simplified
+/// getAbsValue lookup — uses raw map.contains rather than
+/// inVar*Table predicates, which would falsely report neutral
+/// (interval=bottom ∧ addrs=∅) entries as "not present".
+bool AbstractInterpretation::hasAbsValue(const ValVar* var, const ICFGNode* node) const
+{
+    auto it = abstractTrace.find(node);
+    if (it == abstractTrace.end())
+        return false;
+    return it->second.getVarToVal().count(var->getId()) != 0;
+}
+
+bool AbstractInterpretation::hasAbsValue(const ObjVar* var, const ICFGNode* node) const
+{
+    auto it = abstractTrace.find(node);
+    if (it == abstractTrace.end())
+        return false;
+    return it->second.getLocToVal().count(var->getId()) != 0;
+}
+
+bool AbstractInterpretation::hasAbsValue(const SVFVar* var, const ICFGNode* node) const
+{
+    if (const ObjVar* objVar = SVFUtil::dyn_cast<ObjVar>(var))
+        return hasAbsValue(objVar, node);
+    if (const ValVar* valVar = SVFUtil::dyn_cast<ValVar>(var))
+        return hasAbsValue(valVar, node);
+    return false;
+}
+
+void AbstractInterpretation::updateAbsValue(const ValVar* var, const AbstractValue& val, const ICFGNode* node)
+{
+    abstractTrace[node][var->getId()] = val;
+}
+
+void AbstractInterpretation::updateAbsValue(const ObjVar* var, const AbstractValue& val, const ICFGNode* node)
+{
+    AbstractState& as = getAbsState(node);
+    u32_t addr = AbstractState::getVirtualMemAddress(var->getId());
+    as.store(addr, val);
+}
+
+void AbstractInterpretation::updateAbsValue(const SVFVar* var, const AbstractValue& val, const ICFGNode* node)
+{
+    if (const ObjVar* objVar = SVFUtil::dyn_cast<ObjVar>(var))
+        updateAbsValue(objVar, val, node);
+    else if (const ValVar* valVar = SVFUtil::dyn_cast<ValVar>(var))
+        updateAbsValue(valVar, val, node);
+    else
+        assert(false && "Unknown SVFVar kind");
+}
+
+void AbstractInterpretation::getAbsState(const Set<const ValVar*>& vars, AbstractState& result, const ICFGNode* node)
+{
+    AbstractState& as = getAbsState(node);
+    for (const ValVar* var : vars)
+    {
+        u32_t id = var->getId();
+        result[id] = as[id];
+    }
+}
+
+void AbstractInterpretation::getAbsState(const Set<const ObjVar*>& vars, AbstractState& result, const ICFGNode* node)
+{
+    AbstractState& as = getAbsState(node);
+    for (const ObjVar* var : vars)
+    {
+        u32_t addr = AbstractState::getVirtualMemAddress(var->getId());
+        result.store(addr, as.load(addr));
+    }
+}
+
+void AbstractInterpretation::getAbsState(const Set<const SVFVar*>& vars, AbstractState& result, const ICFGNode* node)
+{
+    AbstractState& as = getAbsState(node);
+    for (const SVFVar* var : vars)
+    {
+        if (const ValVar* valVar = SVFUtil::dyn_cast<ValVar>(var))
+        {
+            u32_t id = valVar->getId();
+            result[id] = as[id];
+        }
+        else if (const ObjVar* objVar = SVFUtil::dyn_cast<ObjVar>(var))
+        {
+            u32_t addr = AbstractState::getVirtualMemAddress(objVar->getId());
+            result.store(addr, as.load(addr));
+        }
+    }
+}
+
+IntervalValue AbstractInterpretation::getGepElementIndex(const GepStmt* gep)
+{
+    const ICFGNode* node = gep->getICFGNode();
+    if (gep->isConstantOffset())
+        return IntervalValue((s64_t)gep->accumulateConstantOffset());
+
+    IntervalValue res(0);
+    for (int i = gep->getOffsetVarAndGepTypePairVec().size() - 1; i >= 0; i--)
+    {
+        const ValVar* var = gep->getOffsetVarAndGepTypePairVec()[i].first;
+        const SVFType* type = gep->getOffsetVarAndGepTypePairVec()[i].second;
+
+        s64_t idxLb, idxUb;
+        if (const ConstIntValVar* constInt = SVFUtil::dyn_cast<ConstIntValVar>(var))
+            idxLb = idxUb = constInt->getSExtValue();
+        else
+        {
+            IntervalValue idxItv = getAbsValue(var, node).getInterval();
+            if (idxItv.isBottom())
+                idxLb = idxUb = 0;
+            else
+            {
+                idxLb = idxItv.lb().getIntNumeral();
+                idxUb = idxItv.ub().getIntNumeral();
+            }
+        }
+
+        if (SVFUtil::isa<SVFPointerType>(type))
+        {
+            u32_t elemNum = gep->getAccessPath().getElementNum(gep->getAccessPath().gepSrcPointeeType());
+            idxLb = (double)Options::MaxFieldLimit() / elemNum < idxLb ? Options::MaxFieldLimit() : idxLb * elemNum;
+            idxUb = (double)Options::MaxFieldLimit() / elemNum < idxUb ? Options::MaxFieldLimit() : idxUb * elemNum;
+        }
+        else
+        {
+            if (Options::ModelArrays())
+            {
+                const std::vector<u32_t>& so = PAG::getPAG()->getTypeInfo(type)->getFlattenedElemIdxVec();
+                if (so.empty() || idxUb >= (APOffset)so.size() || idxLb < 0)
+                    idxLb = idxUb = 0;
+                else
+                {
+                    idxLb = PAG::getPAG()->getFlattenedElemIdx(type, idxLb);
+                    idxUb = PAG::getPAG()->getFlattenedElemIdx(type, idxUb);
+                }
+            }
+            else
+                idxLb = idxUb = 0;
+        }
+        res = res + IntervalValue(idxLb, idxUb);
+    }
+    res.meet_with(IntervalValue((s64_t)0, (s64_t)Options::MaxFieldLimit()));
+    if (res.isBottom())
+        res = IntervalValue(0);
+    return res;
+}
+
+IntervalValue AbstractInterpretation::getGepByteOffset(const GepStmt* gep)
+{
+    const ICFGNode* node = gep->getICFGNode();
+    if (gep->isConstantOffset())
+        return IntervalValue((s64_t)gep->accumulateConstantByteOffset());
+
+    IntervalValue res(0);
+    for (int i = gep->getOffsetVarAndGepTypePairVec().size() - 1; i >= 0; i--)
+    {
+        const ValVar* idxOperandVar = gep->getOffsetVarAndGepTypePairVec()[i].first;
+        const SVFType* idxOperandType = gep->getOffsetVarAndGepTypePairVec()[i].second;
+
+        if (SVFUtil::isa<SVFArrayType>(idxOperandType) || SVFUtil::isa<SVFPointerType>(idxOperandType))
+        {
+            u32_t elemByteSize = 1;
+            if (const SVFArrayType* arrOperandType = SVFUtil::dyn_cast<SVFArrayType>(idxOperandType))
+                elemByteSize = arrOperandType->getTypeOfElement()->getByteSize();
+            else if (SVFUtil::isa<SVFPointerType>(idxOperandType))
+                elemByteSize = gep->getAccessPath().gepSrcPointeeType()->getByteSize();
+            else
+                assert(false && "idxOperandType must be ArrType or PtrType");
+
+            if (const ConstIntValVar* op = SVFUtil::dyn_cast<ConstIntValVar>(idxOperandVar))
+            {
+                s64_t lb = (double)Options::MaxFieldLimit() / elemByteSize >= op->getSExtValue()
+                           ? op->getSExtValue() * elemByteSize
+                           : Options::MaxFieldLimit();
+                res = res + IntervalValue(lb, lb);
+            }
+            else
+            {
+                IntervalValue idxVal = getAbsValue(idxOperandVar, node).getInterval();
+                if (idxVal.isBottom())
+                    res = res + IntervalValue(0, 0);
+                else
+                {
+                    s64_t ub = (idxVal.ub().getIntNumeral() < 0) ? 0
+                               : (double)Options::MaxFieldLimit() / elemByteSize >= idxVal.ub().getIntNumeral()
+                               ? elemByteSize * idxVal.ub().getIntNumeral()
+                               : Options::MaxFieldLimit();
+                    s64_t lb = (idxVal.lb().getIntNumeral() < 0) ? 0
+                               : (double)Options::MaxFieldLimit() / elemByteSize >= idxVal.lb().getIntNumeral()
+                               ? elemByteSize * idxVal.lb().getIntNumeral()
+                               : Options::MaxFieldLimit();
+                    res = res + IntervalValue(lb, ub);
+                }
+            }
+        }
+        else if (const SVFStructType* structOperandType = SVFUtil::dyn_cast<SVFStructType>(idxOperandType))
+        {
+            res = res + IntervalValue(gep->getAccessPath().getStructFieldOffset(idxOperandVar, structOperandType));
+        }
+        else
+        {
+            assert(false && "gep type pair only support arr/ptr/struct");
+        }
+    }
+    return res;
+}
+
+AddressValue AbstractInterpretation::getGepObjAddrs(const ValVar* pointer, IntervalValue offset)
+{
+    const ICFGNode* node = pointer->getICFGNode();
+    AddressValue gepAddrs;
+    AbstractState& as = getAbsState(node);
+    APOffset lb = offset.lb().getIntNumeral() < Options::MaxFieldLimit() ? offset.lb().getIntNumeral()
+                  : Options::MaxFieldLimit();
+    APOffset ub = offset.ub().getIntNumeral() < Options::MaxFieldLimit() ? offset.ub().getIntNumeral()
+                  : Options::MaxFieldLimit();
+    for (APOffset i = lb; i <= ub; i++)
+    {
+        const AbstractValue& addrs = getAbsValue(pointer, node);
+        for (const auto& addr : addrs.getAddrs())
+        {
+            s64_t baseObj = as.getIDFromAddr(addr);
+            assert(SVFUtil::isa<ObjVar>(svfir->getSVFVar(baseObj)) && "Fail to get the base object address!");
+            NodeID gepObj = svfir->getGepObjVar(baseObj, i);
+            as[gepObj] = AddressValue(AbstractState::getVirtualMemAddress(gepObj));
+            gepAddrs.insert(AbstractState::getVirtualMemAddress(gepObj));
+        }
+    }
+    return gepAddrs;
+}
+
+AbstractValue AbstractInterpretation::loadValue(const ValVar* pointer, const ICFGNode* node)
+{
+    const AbstractValue& ptrVal = getAbsValue(pointer, node);
+    AbstractState& as = getAbsState(node);
+    AbstractValue res;
+    for (auto addr : ptrVal.getAddrs())
+        res.join_with(as.load(addr));
+    return res;
+}
+
+void AbstractInterpretation::storeValue(const ValVar* pointer, const AbstractValue& val, const ICFGNode* node)
+{
+    const AbstractValue& ptrVal = getAbsValue(pointer, node);
+    AbstractState& as = getAbsState(node);
+    for (auto addr : ptrVal.getAddrs())
+        as.store(addr, val);
+}
+
+const SVFType* AbstractInterpretation::getPointeeElement(const ObjVar* var, const ICFGNode* node)
+{
+    const AbstractValue& ptrVal = getAbsValue(var, node);
+    if (!ptrVal.isAddr())
+        return nullptr;
+    for (auto addr : ptrVal.getAddrs())
+    {
+        NodeID objId = getAbsState(node).getIDFromAddr(addr);
+        if (objId == 0)
+            continue;
+        return svfir->getBaseObject(objId)->getType();
+    }
+    return nullptr;
+}
+
+u32_t AbstractInterpretation::getAllocaInstByteSize(const AddrStmt* addr)
+{
+    const ICFGNode* node = addr->getICFGNode();
+    if (const ObjVar* objvar = SVFUtil::dyn_cast<ObjVar>(addr->getRHSVar()))
+    {
+        if (svfir->getBaseObject(objvar->getId())->isConstantByteSize())
+        {
+            return svfir->getBaseObject(objvar->getId())->getByteSizeOfObj();
+        }
+        else
+        {
+            const std::vector<SVFVar*>& sizes = addr->getArrSize();
+            u32_t elementSize = 1;
+            u64_t res = elementSize;
+            for (const SVFVar* value : sizes)
+            {
+                const AbstractValue& sizeVal = getAbsValue(value, node);
+                IntervalValue itv = sizeVal.getInterval();
+                if (itv.isBottom())
+                    itv = IntervalValue(Options::MaxFieldLimit());
+                res = res * itv.ub().getIntNumeral() > Options::MaxFieldLimit()
+                      ? Options::MaxFieldLimit() : res * itv.ub().getIntNumeral();
+            }
+            return (u32_t)res;
+        }
+    }
+    assert(false && "Addr rhs value is not ObjVar");
+    abort();
+}
+
+Set<const ICFGNode*> AbstractInterpretation::getUseSitesOfObjVar(const ObjVar*, const ICFGNode* node) const
+{
+    Set<const ICFGNode*> succs;
+    for (const auto* edge : node->getOutEdges())
+        succs.insert(edge->getDstNode());
+    return succs;
+}
+
+Set<const ICFGNode*> AbstractInterpretation::getUseSitesOfValVar(const ValVar* var) const
+{
+    Set<const ICFGNode*> succs;
+    if (const ICFGNode* node = var->getICFGNode())
+    {
+        for (const auto* edge : node->getOutEdges())
+            succs.insert(edge->getDstNode());
+    }
+    return succs;
+}
+
+const ICFGNode* AbstractInterpretation::getDefSiteOfValVar(const ValVar* var) const
+{
+    return var->getICFGNode();
+}
+
+const ICFGNode* AbstractInterpretation::getDefSiteOfObjVar(const ObjVar*, const ICFGNode* node) const
+{
+    for (const auto* edge : node->getInEdges())
+        return edge->getSrcNode();
+    return nullptr;
+}
+
+// =====================================================================
+//  Semi-sparse state-access overrides (used by both SemiSparse and
+//  FullSparse subclasses; the latter further restricts ValVar reads).
+// =====================================================================
+
+void SemiSparseAbstractInterpretation::updateAbsState(
+    const ICFGNode* node, const AbstractState& state)
+{
+    // Only replace ObjVar state.  ValVars live at their def-sites and
+    // must not be overwritten when the predecessor's state is merged in.
+    abstractTrace[node].updateAddrStateOnly(state);
+}
+
+void SemiSparseAbstractInterpretation::joinStates(
+    AbstractState& dst, const AbstractState& src)
+{
+    // ValVars live at def-sites in semi-sparse mode; they don't flow
+    // through state merges.  Snapshot dst's ValVar map, perform the full
+    // join, then restore the snapshot — leaving only ObjVars and freed
+    // addrs merged.
+    auto saved = dst.getVarToVal();
+    dst.joinWith(src);
+    dst.clearValVars();
+    for (const auto& [id, val] : saved)
+        dst[id] = val;
+}
+
+const ICFGNode* SemiSparseAbstractInterpretation::getICFGNode(const ValVar* var) const
+{
+    // const ValVars are all defined in global node
+    if (!var->getICFGNode())
+    {
+        return svfir->getICFG()->getGlobalICFGNode();
+    }
+    // for return value of callsite, use the ret-site as def-site
+    else if (SVFUtil::isa<CallICFGNode>(var->getICFGNode()) && SVFUtil::isa<RetValPN>(var))
+    {
+        return SVFUtil::dyn_cast<CallICFGNode>(var->getICFGNode())->getRetICFGNode();
+    }
+    // for other ValVars, use their def-site as the node to query abstract value.
+    else
+    {
+        return var->getICFGNode();
+    }
+}
+
+const AbstractValue& SemiSparseAbstractInterpretation::getAbsValue(
+    const ValVar* var, const ICFGNode* node)
+{
+    return AbstractInterpretation::getAbsValue(var, getICFGNode(var));
+}
+
+bool SemiSparseAbstractInterpretation::hasAbsValue(
+    const ValVar* var, const ICFGNode* node) const
+{
+    return AbstractInterpretation::hasAbsValue(var, getICFGNode(var));
+}
+
+void SemiSparseAbstractInterpretation::updateAbsValue(
+    const ValVar* var, const AbstractValue& val, const ICFGNode* node)
+{
+    // Write to the var's def-site so getAbsValue stays consistent.
+    const ICFGNode* defNode = var->getICFGNode();
+    abstractTrace[defNode ? defNode : node][var->getId()] = val;
+}
+
+// =====================================================================
+//  Full-sparse — SVFG-backed def/use; ValVar reads stubbed.
+// =====================================================================
+
+// TODO(full-sparse): route ValVar reads through the SVFG's reaching-def
+// query.  Stub until that lands so misuse fails loudly instead of
+// silently inheriting semi-sparse semantics.
+const AbstractValue& FullSparseAbstractInterpretation::getAbsValue(
+    const ValVar*, const ICFGNode*)
+{
+    assert(false && "FullSparseAbstractInterpretation::getAbsValue not implemented");
+    abort();
+}
+
+bool FullSparseAbstractInterpretation::hasAbsValue(
+    const ValVar*, const ICFGNode*) const
+{
+    assert(false && "FullSparseAbstractInterpretation::hasAbsValue not implemented");
+    abort();
+}
+
+Set<const ICFGNode*> FullSparseAbstractInterpretation::getUseSitesOfObjVar(
+    const ObjVar* obj, const ICFGNode* node) const
+{
+    assert(svfg && "SVFG is not built for full-sparse AE");
+    return svfg->getUseSitesOfObjVar(obj, node);
+}
+
+Set<const ICFGNode*> FullSparseAbstractInterpretation::getUseSitesOfValVar(
+    const ValVar* var) const
+{
+    assert(svfg && "SVFG is not built for full-sparse AE");
+    return svfg->getUseSitesOfValVar(var);
+}
+
+const ICFGNode* FullSparseAbstractInterpretation::getDefSiteOfValVar(
+    const ValVar* var) const
+{
+    assert(svfg && "SVFG is not built for full-sparse AE");
+    return svfg->getDefSiteOfValVar(var);
+}
+
+const ICFGNode* FullSparseAbstractInterpretation::getDefSiteOfObjVar(
+    const ObjVar* obj, const ICFGNode* node) const
+{
+    assert(svfg && "SVFG is not built for full-sparse AE");
+    return svfg->getDefSiteOfObjVar(obj, node);
+}
